+++ senlin/api/openstack/history.rst
@@ -65,3 +65,5 @@ it can be used by both users and developers.
  - Added ``support_status`` to profile type list.
 
  - Added ``support_status`` to policy type list.
+
+ - Added ``support_status`` to profile type show.